modify rds-s3 clean up script to handle if some of the specified resources does not exist.

+++ tests/e2e/utils/rds-s3/auto-rds-s3-cleanup.py
@@ -1,6 +1,6 @@
 import argparse
 from time import sleep
-import boto3
+import boto3, botocore
 from e2e.fixtures.cluster import delete_iam_service_account
 
 from e2e.utils.utils import (
@@ -27,49 +27,73 @@ def delete_s3_bucket(metadata, secrets_manager_client, region):
     s3_resource = boto3.resource("s3")
     bucket_name = metadata["S3"]["bucket"]
 
-    print("Deleting S3 bucket...")
-
-    bucket = s3_resource.Bucket(bucket_name)
-    bucket.objects.all().delete()
-    s3_client.delete_bucket(Bucket=bucket_name)
+    if check_bucket(bucket_name, s3_client):
+        bucket = s3_resource.Bucket(bucket_name)
+        print("Deleting S3 bucket...")
+        bucket.objects.all().delete()
+        s3_client.delete_bucket(Bucket=bucket_name)
+    else:
+        print("Skip deleting S3 bucket...")
 
     secrets_manager_client.delete_secret(
         SecretId=metadata["S3"]["secretName"], ForceDeleteWithoutRecovery=True
     )
+
+def check_bucket(bucket_name,s3_client):
+    try:
+        s3_client.head_bucket(Bucket=bucket_name)
+        print("Bucket Exists!")
+        return True
+    except botocore.exceptions.ClientError as e:
+        # If a client error is thrown, then check that it was a 404 error.
+        # If it was a 404 error, then the bucket does not exist.
+        error_code = int(e.response['Error']['Code'])
+        if error_code == 403:
+            print("Private Bucket. Forbidden Access!")
+            return True
+        elif error_code == 404:
+            print("Bucket Does Not Exist!")
+            return False
 
 
 def delete_rds(metadata, secrets_manager_client, region):
     rds_client = get_rds_client(region)
     db_instance_name = metadata["RDS"]["instanceName"]
     db_subnet_group_name = metadata["RDS"]["subnetGroupName"]
+    
+    try:
+        rds_client.modify_db_instance(
+            DBInstanceIdentifier=db_instance_name,
+            DeletionProtection=False,
+            ApplyImmediately=True,
+        )
+        print("Deleting RDS instance...")
+        rds_client.delete_db_instance(
+            DBInstanceIdentifier=db_instance_name, SkipFinalSnapshot=True
+        )
+        wait_periods = 30
+        period_length = 30
+        for _ in range(wait_periods):
+            try:
+                if (
+                    rds_client.describe_db_instances(DBInstanceIdentifier=db_instance_name)
+                    is not None
+                ):
+                    sleep(period_length)
+            except:
+                print("RDS instance has been successfully deleted")
+                break
+    except:
+        print("RDS instance does not exist...")
 
-    print("Deleting RDS instance...")
+    
 
-    rds_client.modify_db_instance(
-        DBInstanceIdentifier=db_instance_name,
-        DeletionProtection=False,
-        ApplyImmediately=True,
-    )
-    rds_client.delete_db_instance(
-        DBInstanceIdentifier=db_instance_name, SkipFinalSnapshot=True
-    )
-    wait_periods = 30
-    period_length = 30
-    for _ in range(wait_periods):
-        try:
-            if (
-                rds_client.describe_db_instances(DBInstanceIdentifier=db_instance_name)
-                is not None
-            ):
-                sleep(period_length)
-        except:
-            print("RDS instance has been successfully deleted")
-            break
-
-    print("Deleting DB Subnet Group...")
-
-    rds_client.delete_db_subnet_group(DBSubnetGroupName=db_subnet_group_name)
-    print("DB Subnet Group has been successfully deleted")
+    try:
+        rds_client.delete_db_subnet_group(DBSubnetGroupName=db_subnet_group_name)
+        print("Deleting DB Subnet Group...")
+        print("DB Subnet Group has been successfully deleted")
+    except:
+        print("DB Subnet Group does not exist...")
 
     secrets_manager_client.delete_secret(
         SecretId=metadata["RDS"]["secretName"], ForceDeleteWithoutRecovery=True
